Significant refactor, new launching system
Instead of keeping the existing classpath and blacklisting JARs/classes,
we now replace the entire classpath and load everything ourselves (using
parent = null in the classloader). This means many of the existing hacks
can be removed entirely. This will need a new resolution system to handle
loading vanilla libraries, and a package refactor has been done in
preparation for that.